requirements.txt
asgiref==3.11.0
Django==5.2.8
gunicorn==23.0.0
openpyxl==3.1.5
packaging==25.0
Pillow==12.0.0
psycopg2-binary==2.9.11
pypdf==5.1.0
python-docx==1.1.2
sqlparse==0.5.3
whitenoise==6.8.2
xlrd==2.0.1
